[Build] Continue with the GCC warning silencing, NFC (#22368)
Yet another batch of silencing warnings to make GCC happy.

MODULAR_ORIG_COMMIT_REV_ID: 114af13790ae842dd3a1311aab643dd97a1cc1a9

+++ KGEN/lib/KGENToLLVM/LLVMLoweringUtils.cpp
@@ -1014,9 +1014,9 @@ Value KGEN::convertParameterToLLVM(
       if (!element)
         return {};
 
-      Value destination =
-          b.create<LLVM::GEPOp>(LLVM::LLVMPointerType::get(elementType), ptr,
-                                ArrayRef<LLVM::GEPArg>{idx});
+      Value destination = b.create<LLVM::GEPOp>(
+          LLVM::LLVMPointerType::get(elementType), ptr,
+          ArrayRef<LLVM::GEPArg>{static_cast<int32_t>(idx)});
       b.create<LLVM::StoreOp>(element, destination);
     }
